Rename user_manuals app to man and rely on module navigation
