+++ man/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 
 
-class UserManualsConfig(AppConfig):
+class ManConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "user_manuals"
+    name = "man"
     verbose_name = "3. Protocols - User Manuals"
