+++ pages/templates/admin_doc/index.html
@@ -32,7 +32,7 @@
   <h2><a href="bookmarklets/">{% translate 'Bookmarklets' %}</a></h2>
   <p>{% translate 'Tools for your browser to quickly access admin functionality.' %}</p>
 
-  <h2><a href="{% url 'user_manuals:list' %}">{% translate 'Manuals' %}</a></h2>
+  <h2><a href="{% url 'man:list' %}">{% translate 'Manuals' %}</a></h2>
   <p>{% translate 'Available user manuals and their descriptions.' %}</p>
 </div>
